fix delete food, and enabled debug mode in deployment for easy debug

+++ app/Food.php
@@ -16,7 +16,9 @@ class Food extends Model
 
 
     public function delete(){
+        if($this->image != 'image/logo.png')
         Storage::delete(public_path($this->image));
+        dd($this->image);
         return parent::delete();
 
     }
